apps/code_eval/requirements.txt
tiktoken
chardet
pathspec
GitPython
PyGithub
patool
requests
boto3
pydantic
redis
rq
ollama
ijson
grpcio
grpcio-tools
python-dotenv
openai
google-genai